fetcher: initial, partial implementation
Fetcher takes a directory on disk and assumes a structure like so:
  root - show-name - season-nunmber - episode-name-and-number

For example:
  Vids - Pioneer One - 1 - earthfall-s01e01.mp4

Based on the show names found it will query TvMaze to find the show
information. Then it will proceed to figure out what is already on disk
and eventually write the show to our local store with the appropriate
pending flags.

This work is unfinished.